FYP_NextGenIoTPY/IoTServer/requirements.txt
paho_mqtt==2.1.0
pyserial==3.5
pytz==2022.7.1
